Create extract script that returns expected data.
